+++ UserProfileService/UserProfileService.Application/UseCases/Profile/DeleteProfile/DeleteProfileHandler.cs
@@ -1,4 +1,6 @@
 ﻿using MediatR;
+using UserProfileService.Application.Exceptions;
+using UserProfileService.Domain.Entities;
 using UserProfileService.Domain.Interfaces;
 
 namespace UserProfileService.Application.UseCases.Profile
@@ -17,7 +19,10 @@ namespace UserProfileService.Application.UseCases.Profile
             var profile = await _unitOfWork.ProfileRepository.GetByIdAsync(request.ProfileId);
 
             if (profile == null)
-                throw new KeyNotFoundException("Profile not found.");
+                throw new NotFoundException("Profile not found.");
+
+            if (request.userId != profile!.UserId)
+                throw new UnauthorizedException("Owner isn't valid");
 
             _unitOfWork.ProfileRepository.Delete(profile);
             await _unitOfWork.SaveChangesAsync();
